# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
...
        "/formations": {
            "get": {
                "tags": [
                    "Formation"
                ],
                "summary": "Liste toutes les formations",
                "description": "Point de terminaison pour récupérer la liste de toutes les formations.",
                "operationId": "index()",
                "responses": {
                    "200": {
                        "description": "Liste de toutes les formations",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "formations": [
                                        {
                                            "id": 1,
                                            "nom": "Formation 1"
                                        },
                                        {
                                            "id": 2,
                                            "nom": "Formation 2"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                }
            },
            "post": {
                "tags": [
                    "Formation"
                ],
                "summary": "Enregistre une nouvelle formation",
                "description": "Point de terminaison pour enregistrer une nouvelle formation.",
                "operationId": "store()",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titre": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "date_fin_candidature": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Enregistrement effectué avec succès",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Enregistrement effectué avec succès"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation échouée",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "errors": {
                                        "titre": [
                                            "Le champ titre est requis."
                                        ],
                                        "description": [
                                            "Le champ description est requis."
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Enregistrement échoué"
                    }
                }
            }
        },
        "/formations/{id}": {
            "get": {
                "tags": [
                    "Formation"
                ],
                "summary": "Affiche les détails d'une formation",
                "description": "Point de terminaison pour afficher les détails d'une formation.",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la formation",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la formation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "formations": {
                                        "id": 1,
                                        "titre": "Formation 1",
                                        "description": "Description de la formation",
                                        "date_fin_candidature": "2023-01-01"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Formation non trouvée"
                    },
                    "500": {
                        "description": "Erreur lors de la récupération des détails"
                    }
                }
            },
            "put": {
                "tags": [
                    "Formation"
                ],
                "summary": "Mise à jour d'une formation existante",
                "description": "Point de terminaison pour mettre à jour une formation existante.",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la formation",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titre": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "date_fin_candidature": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mise à jour effectuée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Mise à jour effectuée avec succès"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation échouée",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "errors": {
                                        "titre": [
                                            "Le champ titre est requis."
                                        ],
                                        "description": [
                                            "Le champ description est requis."
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Formation non trouvée"
                    },
                    "500": {
                        "description": "Échec de la mise à jour"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Formation"
                ],
                "summary": "Suppression d'une formation",
                "description": "Point de terminaison pour supprimer une formation existante.",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la formation",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Formation supprimée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "Message": "Supprimé"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Formation non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "Message": "Connais pas"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Échec de la suppression"
                    }
                }
            }
        }
    }
}